Created FLogger implementation and created an Int and Long version of disruption rings and consumers.

+++ src/FortitudeCommon/EventProcessing/Disruption/Consuming/IRingConsumer.cs
@@ -5,7 +5,7 @@ namespace FortitudeCommon.EventProcessing.Disruption.Consuming
 {
     public interface IRingConsumer<T> : IObserver<T> where T : class
     {
-        void OnNext(long sequence, long batchSize, T data, bool startOfBatch, bool endOfBatch);
+        void OnNext(long sequence, int batchSize, T data, bool startOfBatch, bool endOfBatch);
         void OnNext(IEnumerable<T> batch);
     }
 }
