+++ src/FortitudeCommon/EventProcessing/Disruption/Consuming/NoopRingConsumer.cs
@@ -5,7 +5,7 @@ namespace FortitudeCommon.EventProcessing.Disruption.Consuming
 {
     public class NoopRingConsumer<T> : IRingConsumer<T> where T : class
     {
-        public void OnNext(long sequence, long batchSize, T data, bool startOfBatch, bool endOfBatch)
+        public void OnNext(long sequence, int batchSize, T data, bool startOfBatch, bool endOfBatch)
         {
         }
